+++ src/lib/constants.ts
@@ -1,0 +1,3 @@
+export const MIN_LEVEL = 0;
+export const MAX_LEVEL = 100;
+export const DEFAULT_LEVEL = 30;
